Search and before/after functionality
Added Lucene as a dependency to demonstrate using MemoryIndex to score a
tweet for a given query. Works well in Spark so far.

Also added code for determining what users posted between two sets of
tweets. I use this code to determine whether a user is active before and
after an event like the Boston Marathon Bombing.